M    src/main/java/org/picofarad/sdr101/LocalOscillatorSource.java
	*) Removed some casting to make it clearer.


git-svn-id: svn+ssh://ginandtonic.no/home/kms/.svnrepos/sdr101-java@751 bec0a6a6-59ba-0310-824c-86e0bd1f4ed1

+++ src/main/java/org/picofarad/sdr101/LocalOscillatorSource.java
@@ -23,7 +23,7 @@ public class LocalOscillatorSource implements SignalBlock {
 
     public double out() {
 	double r;
-	double position = ((double) sampleNumber) / ((double) sampleRate);
+	double position = (double) sampleNumber / sampleRate;
 	double offset = phaseOffset / 360;
 
 	r = Math.sin(2 * Math.PI * (position + offset) * frequency);
